refactor(api): move AuthenticationKeyExpired error
to identity-access-management bounded context
