+++ api/lib/domain/usecases/reconcile-oidc-user-for-admin.js
@@ -6,7 +6,8 @@
  * @typedef {import ('./index.js').UserLoginRepository} UserLoginRepository
  */
 
-import { AuthenticationKeyExpired, DifferentExternalIdentifierError } from '../errors.js';
+import { AuthenticationKeyExpired } from '../../../src/identity-access-management/domain/errors.js';
+import { DifferentExternalIdentifierError } from '../errors.js';
 import { AuthenticationMethod } from '../models/index.js';
 
 /**
