+++ api/src/identity-access-management/domain/errors.js
@@ -1,5 +1,11 @@
 import { DomainError } from '../../shared/domain/errors.js';
 
+class AuthenticationKeyExpired extends DomainError {
+  constructor(message = 'This authentication key has expired.') {
+    super(message);
+  }
+}
+
 class MissingOrInvalidCredentialsError extends DomainError {
   constructor(message = 'Missing or invalid credentials') {
     super(message);
@@ -19,4 +25,9 @@ class UserShouldChangePasswordError extends DomainError {
   }
 }
 
-export { MissingOrInvalidCredentialsError, PasswordNotMatching, UserShouldChangePasswordError };
+export {
+  AuthenticationKeyExpired,
+  MissingOrInvalidCredentialsError,
+  PasswordNotMatching,
+  UserShouldChangePasswordError,
+};
