+++ api/src/identity-access-management/domain/usecases/create-oidc-user.usecase.js
@@ -1,7 +1,5 @@
-import {
-  AuthenticationKeyExpired,
-  UserAlreadyExistsWithAuthenticationMethodError,
-} from '../../../../lib/domain/errors.js';
+import { UserAlreadyExistsWithAuthenticationMethodError } from '../../../../lib/domain/errors.js';
+import { AuthenticationKeyExpired } from '../errors.js';
 import { UserToCreate } from '../models/UserToCreate.js';
 
 /**
