+++ api/tests/identity-access-management/unit/domain/usecases/create-oidc-user.usecase.test.js
@@ -1,7 +1,5 @@
-import {
-  AuthenticationKeyExpired,
-  UserAlreadyExistsWithAuthenticationMethodError,
-} from '../../../../../lib/domain/errors.js';
+import { UserAlreadyExistsWithAuthenticationMethodError } from '../../../../../lib/domain/errors.js';
+import { AuthenticationKeyExpired } from '../../../../../src/identity-access-management/domain/errors.js';
 import { createOidcUser } from '../../../../../src/identity-access-management/domain/usecases/create-oidc-user.usecase.js';
 import { catchErr, expect, sinon } from '../../../../test-helper.js';
